requirements.txt
Flask==1.1.2
requests==2.25.1
python-dotenv===0.15.0
PyGithub===1.54.1